Move Section component into layout folder
